dbeaver/dbeaver#11684 javadoc and refactoring
