--- plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/impl/sql/DialectQueryGenerator.java
@@ -1,71 +1,0 @@
-/*
- * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2022 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jkiss.dbeaver.model.impl.sql;
-
-import org.jkiss.code.NotNull;
-import org.jkiss.code.Nullable;
-import org.jkiss.dbeaver.model.DBPDataSource;
-import org.jkiss.dbeaver.model.data.DBDAttributeConstraint;
-import org.jkiss.dbeaver.model.data.DBDDataFilter;
-
-import java.util.List;
-
-public interface DialectQueryGenerator {
-
-    void appendQueryConditions(DBPDataSource dataSource,
-        @NotNull StringBuilder query,
-        @Nullable String tableAlias,
-        @Nullable DBDDataFilter dataFilter);
-
-    void appendQueryOrder(DBPDataSource dataSource,
-        @NotNull StringBuilder query,
-        @Nullable String tableAlias,
-        @Nullable DBDDataFilter dataFilter);
-
-    void appendConditionString(@NotNull DBDDataFilter filter,
-        @NotNull DBPDataSource dataSource,
-        @Nullable String conditionTable,
-        @NotNull StringBuilder query,
-        boolean inlineCriteria);
-
-    void appendConditionString(@NotNull DBDDataFilter filter,
-        @NotNull DBPDataSource dataSource,
-        @Nullable String conditionTable,
-        @NotNull StringBuilder query,
-        boolean inlineCriteria,
-        boolean subQuery);
-
-    void appendConditionString(@NotNull DBDDataFilter filter,
-        @NotNull List<DBDAttributeConstraint> constraints,
-        @NotNull DBPDataSource dataSource,
-        @Nullable String conditionTable,
-        @NotNull StringBuilder query,
-        boolean inlineCriteria,
-        boolean subQuery);
-
-    void appendOrderString(@NotNull DBDDataFilter filter,
-        @NotNull DBPDataSource dataSource,
-        @Nullable String conditionTable,
-        boolean subQuery,
-        @NotNull StringBuilder query);
-
-    @Nullable
-    String getConstraintCondition(@NotNull DBPDataSource dataSource,
-        @NotNull DBDAttributeConstraint constraint,
-        @Nullable String conditionTable,
-        boolean inlineCriteria);
-}
